fix(CI): CodeQL/ビルド失敗を修正（N1 の opacity 式の演算子多義性）
CI（Xcode 16 系）で FullPhotoView の `1 - min(pullDown,240)/600` が
"ambiguous use of operator '/'" でビルド失敗（opacity=Double と pullDown=CGFloat の
リテラル型解決が割り算で多義）。CGFloat/Double を明示して解消。ローカル(新しい Xcode)は
通っていたため検出が遅れた。

検証: アプリ iOS ビルド成功。

+++ Packages/PhotoSourceKit/Sources/PhotoSourceKit/Views/FullPhotoView.swift
@@ -37,8 +37,9 @@ struct FullPhotoView<Store: PhotoStore>: View {
                     photo
                         .frame(width: geo.size.width, height: geo.size.height)
                         // N1: 引っ張りに応じて軽く縮小＋退色させ「閉じる」フィードバックを出す。
-                        .scaleEffect(1 - min(pullDown, 240) / 2400, anchor: .center)
-                        .opacity(1 - min(pullDown, 240) / 600)
+                        // 型は明示（CGFloat/Double を混ぜると割り算演算子が多義になりビルドが落ちる）。
+                        .scaleEffect(1 - min(pullDown, CGFloat(240)) / CGFloat(2400), anchor: .center)
+                        .opacity(1 - Double(min(pullDown, 240)) / 600)
                     PhotoInfoPanel(
                         captureDate: item.captureDate,
                         placeName: placeName,
